requirements.txt
annotated-types==0.6.0
anyio==4.3.0
assemblyai==0.26.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.110.2
ffmpeg==1.4
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
nodeenv==1.8.0
packaging==24.0
postgrest==0.16.3
prisma==0.13.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.4
requests==2.32.1
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
tomlkit==0.12.4
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
